Made separate header files metadata struct declarations.
