+++ include/data/data_sprites.h
@@ -1,0 +1,9 @@
+// data/data_sprites.h
+#pragma once
+#include <raylib.h>
+#include <string>
+
+struct SpriteMetaData {
+  std::string name;
+  Texture *sprite;
+};
