+++ src/system/audio.cpp
@@ -12,20 +12,6 @@
 using std::string, std::tuple, std::make_tuple, std::tie;
 
 
-SoundMetaData::SoundMetaData(string name, Sound *sound, bool random_pitch,
-                             float min_pitch, float max_pitch) 
-{
-  this->name = name;
-  this->sound = sound;
-
-  this->random_pitch = random_pitch;
-  this->min_pitch = min_pitch;
-  this->max_pitch = max_pitch;
-
-  PLOGD << "Saved sound data for: " << name;
-}
-
-
 AudioManager::AudioManager() {
   meta_data = toml::parse("audio/audio_data.toml");
   loadSoundEffects();
@@ -70,8 +56,7 @@ void AudioManager::loadSoundEffects() {
     Sound *sound_ptr = &sound_effects[index];
 
     audio::sfx_metadata.push_back(
-      SoundMetaData(sound_name, sound_ptr, use_random_pitch, 
-                    min_pitch, max_pitch)
+      {sound_name, sound_ptr, use_random_pitch, min_pitch, max_pitch}
     );
   }
